fixup! Add Cube controller events
