+++ docker/src/test/java/org/arquillian/cube/impl/client/CubeLifecycleControllerTest.java
@@ -8,10 +8,10 @@ import java.util.List;
 import org.arquillian.cube.impl.model.DockerCubeRegistry;
 import org.arquillian.cube.spi.Cube;
 import org.arquillian.cube.spi.CubeRegistry;
-import org.arquillian.cube.spi.events.CreateCube;
-import org.arquillian.cube.spi.events.DestroyCube;
-import org.arquillian.cube.spi.events.StartCube;
-import org.arquillian.cube.spi.events.StopCube;
+import org.arquillian.cube.spi.event.CreateCube;
+import org.arquillian.cube.spi.event.DestroyCube;
+import org.arquillian.cube.spi.event.StartCube;
+import org.arquillian.cube.spi.event.StopCube;
 import org.jboss.arquillian.core.api.annotation.ApplicationScoped;
 import org.jboss.arquillian.core.test.AbstractManagerTestBase;
 import org.junit.Before;
