+++ docker/src/test/java/org/arquillian/cube/impl/client/container/CubeContainerLifecycleControllerTest.java
@@ -8,10 +8,10 @@ import java.util.List;
 import org.arquillian.cube.impl.model.DockerCubeRegistry;
 import org.arquillian.cube.spi.Cube;
 import org.arquillian.cube.spi.CubeRegistry;
-import org.arquillian.cube.spi.events.CreateCube;
-import org.arquillian.cube.spi.events.DestroyCube;
-import org.arquillian.cube.spi.events.StartCube;
-import org.arquillian.cube.spi.events.StopCube;
+import org.arquillian.cube.spi.event.CreateCube;
+import org.arquillian.cube.spi.event.DestroyCube;
+import org.arquillian.cube.spi.event.StartCube;
+import org.arquillian.cube.spi.event.StopCube;
 import org.jboss.arquillian.container.spi.Container;
 import org.jboss.arquillian.container.spi.ContainerRegistry;
 import org.jboss.arquillian.container.spi.client.container.DeployableContainer;
