+++ spi/src/main/java/org/arquillian/cube/spi/event/CreateCube.java
@@ -1,16 +1,16 @@
-package org.arquillian.cube.spi.events;
+package org.arquillian.cube.spi.event;
 
 import org.arquillian.cube.spi.Cube;
 
-public class StopCube implements CubeControlEvent {
+public class CreateCube implements CubeControlEvent {
 
     private String cubeId;
 
-    public StopCube(Cube cube) {
+    public CreateCube(Cube cube) {
         this(cube.getId());
     }
 
-    public StopCube(String cubeId) {
+    public CreateCube(String cubeId) {
         this.cubeId = cubeId;
     }
