+++ spi/src/main/java/org/arquillian/cube/spi/event/StopCube.java
@@ -1,16 +1,16 @@
-package org.arquillian.cube.spi.events;
+package org.arquillian.cube.spi.event;
 
 import org.arquillian.cube.spi.Cube;
 
-public class DestroyCube implements CubeControlEvent {
+public class StopCube implements CubeControlEvent {
 
     private String cubeId;
 
-    public DestroyCube(Cube cube) {
+    public StopCube(Cube cube) {
         this(cube.getId());
     }
 
-    public DestroyCube(String cubeId) {
+    public StopCube(String cubeId) {
         this.cubeId = cubeId;
     }
